+++ load/tpcc/payment.cpp
@@ -21,7 +21,7 @@ int payment(int w_id_arg, int d_id_arg, int byname,
 //              << (void *)&t << ", function [payment]@"
 //              << (void *)payment << std::endl;
     transaction_obj tran_obj;
-    std::string tran_name = parseInt((int)t) + "@payment_" + parseInt(t_id);
+    std::string tran_name = parseInt((long)t) + "@payment_" + parseInt(t_id);
     int w_id = w_id_arg;
     int d_id = d_id_arg;
     int c_id = c_id_arg;
